Remove unused and non-necessary items
